group agile into stack
